Split ParallelFor and ParallelReduce OpenACC specializations
